fix test kenlm_path type

+++ test.py
@@ -7,7 +7,7 @@ if __name__ == '__main__':
     parser.add_argument('-f', '--test_meta_filepath', type=str, required=True, help='test meta file path')
     parser.add_argument('-m', '--model_path', type=str, required=True, help='model path')
     parser.add_argument('-b', '--beam_size', type=int, default=1, help='beam size for ctc beamseach decoder, 1 mean greedy decode')
-    parser.add_argument('-l', '--kenlm_path', type=int, help='kenlm model for ctc beamseach decoder')
+    parser.add_argument('-l', '--kenlm_path', type=float, help='kenlm model for ctc beamseach decoder')
     parser.add_argument('--kenlm_alpha', type=float, default=0.2, help='kenlm alpha for ctc beamseach decoder')
     parser.add_argument('--kenlm_beta', type=float, default=1.5, help='kenlm beta for ctc beamseach decoder')
     parser.add_argument('-d', '--device', type=str, default="cpu", help='test on device')
